Move SaveImagesModel from ImagingImms to Imaging directory
Move SaveImagesViewModel from ImagingImms to Imaging directory
