+++ src/MSDIAL5/MsdialGuiApp/ViewModel/Imaging/SaveImagesViewModel.cs
@@ -1,4 +1,4 @@
-﻿using CompMs.App.Msdial.Model.ImagingImms;
+﻿using CompMs.App.Msdial.Model.Imaging;
 using CompMs.App.Msdial.ViewModel.Service;
 using CompMs.CommonMVVM;
 using Reactive.Bindings;
@@ -6,7 +6,7 @@ using Reactive.Bindings.Extensions;
 using Reactive.Bindings.Notifiers;
 using System;
 
-namespace CompMs.App.Msdial.ViewModel.ImagingImms
+namespace CompMs.App.Msdial.ViewModel.Imaging
 {
     internal sealed class SaveImagesViewModel : ViewModelBase
     {
@@ -19,7 +19,7 @@ namespace CompMs.App.Msdial.ViewModel.ImagingImms
             _broker = broker ?? throw new ArgumentNullException(nameof(broker));
             Path = model.ToReactivePropertyAsSynchronized(m => m.Path).AddTo(Disposables);
             BrowseCommand = new ReactiveCommand().WithSubscribe(Browse).AddTo(Disposables);
-            SaveCommand = new AsyncReactiveCommand().WithSubscribe(model.SaveAsync).AddTo(Disposables);
+            SaveCommand = new AsyncReactiveCommand().WithSubscribe(() => model.SaveAsync()).AddTo(Disposables);
         }
 
         public ReactiveProperty<string> Path { get; }
@@ -31,7 +31,7 @@ namespace CompMs.App.Msdial.ViewModel.ImagingImms
         {
             var request = new SaveFileNameRequest(path => Path.Value = path)
             {
-                Filter = string.Join("|", new[] { "png|.png", "gif|.gif" }),
+                Filter = string.Join("|", ["png|.png", "gif|.gif"]),
             };
             _broker.Publish(request);
         }
